apply ts on IngredientConditionDTO
